Reuse service extraction in c2c tooling
