+++ c2c-extension-test/src/test/resources/testng-integration.xml
@@ -35,14 +35,9 @@
             <class name="io.ballerina.c2c.test.samples.DockerSingleTest"/>
             <class name="io.ballerina.c2c.test.SettingsTest"/>
             <class name="io.ballerina.c2c.test.MultipleConfigTest"/>
-        </classes>
-    </test>
-
-    <test name="ballerina-kubernetes-tests" parallel="false">
-        <classes>
+            <class name="io.ballerina.c2c.test.ServiceExtractionTest"/>
             <class name="io.ballerina.c2c.test.JobTest"/>
             <class name="io.ballerina.c2c.test.EnvTest"/>
-            <class name="io.ballerina.c2c.test.ServiceExtractionTest"/>
             <class name="io.ballerina.c2c.test.CustomDiagnosticsTest"/>
             <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidServicePath"/>
             <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidResourcePath"/>
@@ -50,12 +45,10 @@
             <class name="io.ballerina.c2c.test.codeactions.CloudEnvTest"/>
             <class name="io.ballerina.c2c.test.codeactions.CloudProbesTest"/>
             <class name="io.ballerina.c2c.test.completion.CloudTest"/>
+            <class name="io.ballerina.c2c.test.completion.BallerinaTest"/>
             <class name="io.ballerina.c2c.test.command.CreateCloudTomlCommandTest"/>
             <class name="io.ballerina.c2c.test.docker.CmdTest"/>
             <class name="io.ballerina.c2c.test.ServiceTest"/>
-            <class name="io.ballerina.c2c.test.SettingsTest"/>
-            <class name="io.ballerina.c2c.test.MultipleConfigTest"/>
         </classes>
     </test>
-
 </suite>
