+++ c2c-tooling/src/main/java/io/ballerina/c2c/tooling/codeaction/providers/kubernetes/AbstractInvalidResourceCodeAction.java
@@ -17,9 +17,9 @@
  */
 package io.ballerina.c2c.tooling.codeaction.providers.kubernetes;
 
-import io.ballerina.c2c.tooling.codeaction.toml.ProjectServiceInfo;
+import io.ballerina.c2c.diagnostics.ProjectServiceInfo;
+import io.ballerina.c2c.diagnostics.ServiceInfo;
 import io.ballerina.c2c.tooling.codeaction.toml.ProjectServiceInfoHolder;
-import io.ballerina.c2c.tooling.codeaction.toml.ServiceInfo;
 import io.ballerina.c2c.tooling.toml.CommonUtil;
 import io.ballerina.c2c.tooling.toml.Probe;
 import io.ballerina.c2c.tooling.toml.TomlSyntaxTreeUtil;
@@ -42,7 +42,6 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 
 /**
@@ -56,43 +55,41 @@ public abstract class AbstractInvalidResourceCodeAction extends ProbeBasedDiagno
         Optional<Project> project = ctx.workspace().project(ctx.filePath());
         ProjectServiceInfo projectServiceInfo =
                 ProjectServiceInfoHolder.getInstance(ctx.languageServercontext()).getProjectInfo(project.orElseThrow());
-        Map<String, List<ServiceInfo>> serviceList = projectServiceInfo.getServiceMap();
+        List<ServiceInfo> serviceList = projectServiceInfo.getServiceList();
         List<CodeAction> codeActionList = new ArrayList<>();
-        for (Map.Entry<String, List<ServiceInfo>> entry : serviceList.entrySet()) {
-            String filePath = entry.getKey();
-            List<ServiceInfo> serviceInfos = entry.getValue();
-            for (ServiceInfo service : serviceInfos) {
-                int port = service.getListener().getPort();
-                if (probe.getPort().getValue() == port) {
-                    Path balFilePath = ctx.workspace().projectRoot(ctx.filePath()).resolve(filePath);
-                    NodeList<Node> members = service.getNode().members();
-                    Node lastResource = members.get(members.size() - 1);
-                    Position position = new Position(lastResource.lineRange().endLine().line() + 1, 0);
+        for (ServiceInfo service : serviceList) {
+            String filePath = service.getNode().syntaxTree().filePath();
+            int port = service.getListener().getPort();
+            if (probe.getPort().getValue() == port) {
+                Path balFilePath = ctx.workspace().projectRoot(ctx.filePath()).resolve(filePath);
+                NodeList<Node> members = service.getNode().members();
+                Node lastResource = members.get(members.size() - 1);
+                Position position = new Position(lastResource.lineRange().endLine().line() + 1, 0);
 
-                    CodeAction action = new CodeAction();
-                    action.setKind(CodeActionKind.QuickFix);
-                    String importText = generateProbeFunctionText(service, probe);
-                    List<TextEdit> edits = Collections.singletonList(
-                            new TextEdit(new Range(position, position), importText));
-                    action.setEdit(new WorkspaceEdit(Collections.singletonList(Either.forLeft(
-                            new TextDocumentEdit(
-                                    new VersionedTextDocumentIdentifier(balFilePath.toUri().toString(), null),
-                                    edits)))));
-                    action.setTitle("Add Resource to Service");
-                    List<Diagnostic> cursorDiagnostics = new ArrayList<>();
-                    cursorDiagnostics.add(diagnostic);
-                    action.setDiagnostics(cursorDiagnostics);
-                    codeActionList.add(action);
-                    break;
-                }
+                CodeAction action = new CodeAction();
+                action.setKind(CodeActionKind.QuickFix);
+                String importText = generateProbeFunctionText(service, probe);
+                List<TextEdit> edits = Collections.singletonList(
+                        new TextEdit(new Range(position, position), importText));
+                action.setEdit(new WorkspaceEdit(Collections.singletonList(Either.forLeft(
+                        new TextDocumentEdit(
+                                new VersionedTextDocumentIdentifier(balFilePath.toUri().toString(), null),
+                                edits)))));
+                action.setTitle("Add Resource to Service");
+                List<Diagnostic> cursorDiagnostics = new ArrayList<>();
+                cursorDiagnostics.add(diagnostic);
+                action.setDiagnostics(cursorDiagnostics);
+                codeActionList.add(action);
+                break;
             }
         }
+
         return codeActionList;
     }
 
     private String generateProbeFunctionText(ServiceInfo service, Probe probe) {
         String tomlPath = TomlSyntaxTreeUtil.trimResourcePath(probe.getPath().getValue());
-        String serviceName = TomlSyntaxTreeUtil.trimResourcePath(service.getServiceName());
+        String serviceName = TomlSyntaxTreeUtil.trimResourcePath(service.getServicePath());
         String serviceResourcePath;
         if (tomlPath.equals(serviceName)) {
             serviceResourcePath = ".";
