--- c2c-tooling/src/main/java/io/ballerina/c2c/tooling/codeaction/toml/ListenerInfo.java
@@ -1,55 +1,0 @@
-/*
- * Copyright (c) 2021, WSO2 Inc. (http://wso2.com) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.ballerina.c2c.tooling.codeaction.toml;
-
-/**
- * Represents Listener information of a ballerina document.
- *
- * @since 2.0.0
- */
-public class ListenerInfo {
-    private String name;
-    private int port;
-
-    public ListenerInfo(String name, int port) {
-        this.name = name;
-        this.port = port;
-    }
-
-    public int getPort() {
-        return port;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public void setPort(int port) {
-        this.port = port;
-    }
-
-    @Override
-    public String toString() {
-        return "ListenerInfo{" +
-                "name='" + name + '\'' +
-                ", port=" + port +
-                '}';
-    }
-}
